refactor: split useWalletBoundReferralCode into modular structure
Split the monolithic useWalletBoundReferralCode.tsx into separate modules:
- InviteCodeDialog.tsx: Dialog component for invite code input
- useGetReferralCodeWalletInfo.ts: Hook for fetching wallet info
- useWalletBoundReferralCode.tsx: Main hook with bind logic
- types.ts: Shared type definitions
- index.ts: Module exports
