+++ packages/kit/src/views/ReferFriends/hooks/useWalletBoundReferralCode/index.ts
@@ -1,0 +1,4 @@
+export { useWalletBoundReferralCode } from './useWalletBoundReferralCode';
+export { useGetReferralCodeWalletInfo } from './useGetReferralCodeWalletInfo';
+export { InviteCodeDialog } from './InviteCodeDialog';
+export type { IReferralCodeWalletInfo } from './types';
